+++ uniffi_bindgen/src/bindings/swift/mod.rs
@@ -30,8 +30,9 @@
 //!
 
 use crate::{
-    bindings::GenerateOptions, interface::rename, BindgenLoader, BindgenPaths, Component,
-    ComponentInterface,
+    bindings::GenerateOptions,
+    interface::{apply_exclusions, rename},
+    BindgenLoader, BindgenPaths, Component, ComponentInterface,
 };
 use anyhow::{bail, Result};
 use camino::Utf8PathBuf;
@@ -229,6 +230,11 @@ pub struct SwiftBindingsOptions {
 
 // A helper for renaming items.
 fn apply_renames(components: &mut Vec<Component<Config>>) {
+    // Remove excluded items, this happens before renaming
+    for c in components.iter_mut() {
+        apply_exclusions(&mut c.ci, &c.config.exclude);
+    }
+
     let mut module_renames = HashMap::new();
     // Collect all rename configurations from all components, keyed by module_path
     for c in components.iter() {